summarizer/requirements.txt
# Dipendenze minimali per il sistema di summarizzazione
openai>=1.0.0
python-dotenv>=1.0.0